requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
blinker==1.4
Brlapi==0.7.0
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
confusable-homoglyphs==3.2.0
cryptography==3.2.1
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.4
django-allauth==0.44.0
django-cors-headers==3.6.0
django-registration==3.1.1
django-rest-auth==0.9.5
djangorestframework==3.12.2
entrypoints==0.3
evdev==1.3.0
filelock==3.0.12
greenlet==0.4.15
hidpidaemon==18.4.6
httplib2==0.14.0
idna==2.10
isort==5.7.0
jedi==0.18.0
kernelstub==3.1.4
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.12.0
macaroonbakery==1.3.1
mccabe==0.6.1
msgpack==0.6.2
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
parso==0.8.1
Pillow==8.0.1
pop-transition==1.0.1
protobuf==3.6.1
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.20
pycups==1.9.73
pydbus==0.6.0
PyGObject==3.36.0
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pynvim==0.4.1
pyRFC3339==1.1
python-apt===2.1.2pop0-1587756471-20.04-cd2988e
python-debian===0.1.36ubuntu1
python-distutils-extra==2.39
python-Levenshtein==0.12.0
python-magic==0.4.16
python-xlib==0.23
python3-openid==3.2.0
pytz==2020.4
pyudev==0.21.0
pyxdg==0.26
PyYAML==5.3.1
repoman==1.2.2
requests==2.25.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
sessioninstaller==0.0.0
setproctitle==1.1.10
simplejson==3.16.0
six==1.15.0
sqlparse==0.4.1
systemd-python==234
toml==0.10.2
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
ulauncher==5.9.0
urllib3==1.26.2
virtualenv==20.2.2
wadllib==1.3.3
websocket-client==0.53.0
wrapt==1.12.1
xkit==0.0.0
yapf==0.30.0
